% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{captureSemantics}
\alias{captureSemantics}
\title{Show whether a function is wrapped or not}
\usage{
captureSemantics(fun)
}
\arguments{
\item{fun}{The function that is / is not wrapped}
}
\value{
TRUE/FALSE, depending on whether the function is wrapped
}
\description{
Show whether a function is wrapped or not
}
\seealso{
\code{\link{captureSemantics<-}}
}

