% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{disableProvenance}
\alias{disableProvenance}
\title{Enable / disable provenance tracking}
\usage{
disableProvenance()
}
\description{
Enable / disable provenance tracking
}
\seealso{
\code{\link{enableProvenance}}, {\link{reset_Provenance}}
}

