% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{enableProvenance}
\alias{enableProvenance}
\title{Enable / disable provenance tracking}
\usage{
enableProvenance()
}
\value{
(invisible)
}
\description{
Enable / disable provenance tracking
}
\seealso{
\code{\link{disableProvenance}}, \code{\link{reset_provenance}}
}

