% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{estimateSemantics}
\alias{estimateSemantics}
\title{Estimate semantics of a given object}
\usage{
estimateSemantics(var, env = globalenv(), isLiteral = FALSE)
}
\arguments{
\item{isLiteral}{}
}
\description{
Estimate semantics of a given object
}

