% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{functionalType}
\alias{functionalType}
\title{Get functional type of a spatio-temporal dataset}
\usage{
functionalType(obj, attr = "ALL")
}
\arguments{
\item{obj}{The object from which the functional type shall be retrieved}

\item{attr}{The attribute from which the functional type shall be retrieved}
}
\value{
The object that is annotated
}
\description{
Get functional type of a spatio-temporal dataset
}
\seealso{
\code{\link{functionalType<-}}
}

