% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getCallSemantics}
\alias{getCallSemantics}
\title{Pre-defined semantics of a function call}
\usage{
getCallSemantics(x)
}
\arguments{
\item{x}{An object of type function}
}
\value{
- returns a list of pre-defined semantics which are alowed for this function
- returns "dynamic" if the function is semantics-enabled but does not have pre-defined semantics
- returns NULL if the function is not semantics-enabled and does not implement pre-defined semantics
}
\description{
Pre-defined semantics of a function call
}
\seealso{
\code{\link{getObjectSemantics}}, \code{\link{captureSemantics}}
}

