% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getDefaultCallSemantics}
\alias{getDefaultCallSemantics}
\title{Get default call semantics that where defined for a wrapped function}
\usage{
getDefaultCallSemantics(x)
}
\arguments{
\item{x}{An object of type function, should be a semantic wrapper function}
}
\value{
- returns a list of pre-defined semantics which are alowed for this function
- returns "dynamic" if the function is semantics-enabled but does not have pre-defined semantics
- returns NULL if the function is not semantics-enabled and does not implement pre-defined semantics
}
\description{
Get default call semantics that where defined for a wrapped function
}
\seealso{
\code{\link{getObjectSemantics}}, \code{\link{captureSemantics}}
}

