% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getObjectSemantics}
\alias{getObjectSemantics}
\title{Estimate semantics of a given object}
\usage{
getObjectSemantics(var, env = globalenv(), isLiteral = FALSE)
}
\arguments{
\item{var}{object or name of the object}

\item{env}{The environment in which the object shall be evaluated}

\item{isLiteral}{TRUE/FALSE allowed. Force var to be interpreted as a literal, even if it is for instance a name of an object in the workspace.}
}
\value{
Returns the semantic reference type of the object.
 If the reference type is uncertain because a required annotation is missing, a reference type is assumed and prefixed with a questionmark (?), also a warning is given out.
 For objects, where no semantic mapping is defined, simply the class will be returned
}
\description{
Estimate semantics of a given object
}

