% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getVersions}
\alias{getVersions}
\title{Get history of a variable's name bindings}
\usage{
getVersions(var)
}
\arguments{
\item{var}{Name of the variable, either string or object}
}
\value{
A data.frame describing changes of a variable's name binding
}
\description{
This function returns a versioning history showing all modifications (initialization, updates and replacements) of a variable's name binding in the global environment
The version based on recording provenance, i.e. it only captures modifcations that where don after calling enableProvenance() and before calling disableProvenance()
}

