% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{reset_provenance}
\alias{reset_provenance}
\title{Reset / delete internal provenance record}
\usage{
reset_provenance()
}
\value{
(invisible)
}
\description{
Deletes all internally recorded provenance and sets the tracker back to default state
}
\seealso{
\code{\link{enableProvenance}}, \code{\link{disableProvenance}}
}

