% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{spsem_internals}
\alias{spsem_internals}
\title{Retrieves packages internals}
\usage{
spsem_internals()
}
\value{
An environment in which all internal variables and functions of the SpatialSemantics-package are saved
}
\description{
Retrieves packages internals
}

