% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{addSemanticPedigree}
\alias{addSemanticPedigree}
\title{Add semantic pedigree}
\usage{
addSemanticPedigree(obj, attr = "ALL", name = NA, procedure,
  result_semantics = NULL, parent_semantics = NULL)
}
\arguments{
\item{parent_semantics}{}
}
\description{
Add semantic pedigree
}

