% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{captureSemantics<-}
\alias{captureSemantics<-}
\title{Capture Semantics}
\usage{
captureSemantics(fun, semantics = NA, procedureName = "unknown",
  validator = NULL,
  postprocessor = algebr$genericProcedureAnnotator(procedureName)) <- value
}
\arguments{
\item{value}{}
}
\description{
Capture Semantics
}

