% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{disableProvenance}
\alias{disableProvenance}
\title{Enable / disable provenance tracking}
\usage{
disableProvenance()
}
\value{
(invisible)
}
\description{
Enable / disable provenance tracking
}
\seealso{
\code{\link{enableProvenance}}, \code{\link{reset_provenance}}
}

