% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{functionalType<-}
\alias{functionalType<-}
\title{Add a functional type to semantic pedigree}
\usage{
functionalType(obj, attr = "ALL") <- value
}
\arguments{
\item{value}{}
}
\description{
Add a functional type to semantic pedigree
}

