% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{functionalType}
\alias{functionalType}
\title{Get functional type from semantic pedigree}
\usage{
functionalType(obj, attr = "ALL")
}
\arguments{
\item{obj}{The object from which the functional type shall be retrieved}

\item{attr}{The attribtute from which the functional type shall be retrieved}
}
\description{
Get functional type from semantic pedigree
}

