% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getScriptGraph}
\alias{getScriptGraph}
\title{Retrieve derivation graph for export / vizalization}
\usage{
getScriptGraph(g = algebr$scriptGraph)
}
\arguments{
\item{g}{Derivation graph in internal data.frame format as created by algebr$newDerivationGraph}
}
\description{
Returns derivation graph of the script in Ragraph format (Rgraphviz)
can also convert other graphs created by algebr$newDerivationGraph to the Ragraph format (g-parameter)
}

