% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getScriptGraph}
\alias{getScriptGraph}
\title{Retrieve derivation graph for export / vizalization}
\usage{
getScriptGraph(g = algebr$scriptGraph)
}
\arguments{
\item{g}{Derivation graph in internal data.frame format as created by the internal function 'algebr$newDerivationGraph'}
}
\value{
Returns derivation graph of the script in Ragraph format (Rgraphviz) or NULL, if there is no provanance available
}
\description{
This function returns a derivation graph from the previously recorded provenance, if any
}

