% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getVersions}
\alias{getVersions}
\title{Version history of variable bindings}
\usage{
getVersions(var)
}
\arguments{
\item{var}{name of the variable, either string or object}
}
\description{
This function returns a versioning history showing all modifications (initialization, updates and replacements) of a particular variable in the global environment
The version based on recording provenance, i.e. it only captures modifcations that where don after calling enableProvenance() and before calling disableProvenance()
}

