% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{reset_provenance}
\alias{reset_provenance}
\title{Deletes all internally recorded provenance and sets the tracker back to default state}
\usage{
reset_provenance()
}
\description{
Deletes all internally recorded provenance and sets the tracker back to default state
}

